Add files via upload
In this exercise I had complete the code of a simulator (written in the C language) of an architecture
composed of a very simple CPU and a memory. This includes, already implemented, a console where it
is possible for a user to give commands to read and write to the central memory, including reading to
memory the values represented in a text file. In the repo you have all the code(p.code, dorun.c, sim.c) and the PDFs in portuguese and english version